Refactor : 마크다운 에디터 상태 분리
