+++ front/src/components/commons/MarkdownEditor/index.js
@@ -1,11 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import CustomTextarea from './CustomTextarea';
 import UploadInput from './UploadInput';
 import ContentsLengthLayer from './ContentsLengthLayer';
 import color from '../../../libs/color';
-
-const DELAY = 2000;
 
 const EditorContainer = styled.div`
   width: 100%;
@@ -24,38 +22,11 @@ const EditorInputContainer = styled.div`
   }
 `;
 
-let debounceTimer;
-
-const MarkdownEditor = ({ onChange }) => {
-  const [contentsLength, setContentsLength] = useState(0);
-  const [showContentsLength, setShowContentsLength] = useState(false);
-
-  const setHideContentsLengthTimer = () => {
-    setTimeout(() => {
-      setShowContentsLength(false);
-    }, DELAY);
-  };
-
-  const setShowContentsLengthDebounce = (newContentsLength) => {
-    if (debounceTimer) {
-      clearTimeout(debounceTimer);
-    }
-    debounceTimer = setTimeout(() => {
-      setShowContentsLength(true);
-      setContentsLength(newContentsLength);
-      setHideContentsLengthTimer();
-    }, DELAY);
-  };
-
-  const updateContents = (newContents) => {
-    onChange(newContents);
-    setShowContentsLengthDebounce(newContents.length);
-  };
-
+const MarkdownEditor = ({ contentsLength, showContentsLength, onChange }) => {
   return (
     <EditorContainer>
       <EditorInputContainer>
-        <CustomTextarea onChange={updateContents} />
+        <CustomTextarea onChange={onChange} />
         {showContentsLength && (
           <ContentsLengthLayer
             className="markdown-editor__content-length"
